+++ client/src/modules/Course/Styles/AnonymousWarning.module.css
@@ -9,7 +9,7 @@
   border-radius: 8px;
 }
 
-.anonymousLogo {
+.emoji {
   max-width: 25%;
   max-height: 25%;
   margin: auto;
